Major overhaul to mzParser to allow for multiple possible precursor charge states to be assigned in the RAMP interface. This is extended, but not tested, for multiple precursor ions (think MSX) with multiple possible precursor charges.
Please note that MSReader in the MSToolkit does not know how to handle multiple precursors when reading mzML files. Also note that there may need to be additional changes when dealing with "selected ion m/z" fields in msconvert versions of mzML, which may be the isolated mass (equivalent to "isolation window target m/z") or the true monoisotopic mass (equivalent to "[Thermo Trailer Extra]Monoisotopic M/Z:"). There is no indication which one from the tag itself, and thus MSReader prefers to use the Thermo Trailer Extra value, and in its absence, will default to the selected ion m/z value. In applications, such as Comet, which rely on the Z lines, there is no indication which value was obtained.